requirements.txt
djangorestframework==3.10.0
pyPDF2==1.26.0